+++ CHANGELOG.md
@@ -10,10 +10,14 @@ Copy and pasting the git commit messages is __NOT__ enough.
 
 ## [Unreleased]
 ### Added
+- Added action formatting option to `hiway-v0`.
 ### Changed
+- Moved action and observation conversions from `smarts.env.gymnasium.utils` to `smarts.env.utils`.
 ### Deprecated
 ### Fixed
 - Fixed an issue where metrics break down with unformatted observations.
+- Fixed an issue where `hiway-v1` would cause an exception when using "unformatted" observations.
+- Unformatted actions an observations in `hiway-v0` provide `None` rather than an incorrect space.
 ### Removed
 ### Security